added post routes with passport local strategy authentication
