+++ controllers/api/auth.js
@@ -1,0 +1,39 @@
+const passport = require('passport');
+const LocalStrategy = require('passport-local');
+const express = require('express');
+const router = express.Router();
+
+passport.use(new LocalStrategy(
+  function(username, password, done) {
+    User.findOne({ username: username }, function (err, user) {
+      if (err) { return done(err); }
+      if (!user) { return done(null, false); }
+      if (!user.checkPassword(password)) { return done(null, false); }
+      return done(null, user);
+    });
+  }
+));
+
+passport.serializeUser((user, done) => {
+  process.nextTick(() => {
+    done(null, {id: user.id, username: user.username});
+  });
+});
+
+passport.deserializeUser((user, done) => {
+  process.nextTick(() => {
+    return done(null, user);
+  });
+});
+
+router.post('/login', (req, res, next) => {
+  res.render('login');
+});
+
+router.post('/login', passport.authenticate('local', {
+  successReturnToOrRedirect: '/',
+  failureRedirect: '/login',
+  failureMessage: true
+}));
+
+module.exports = router;
